LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/memory_init64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (memory_init64.wast:6) assert_equals: expected false but got true
FAIL #3 Test that WebAssembly compilation succeeds (memory_init64.wast:50) assert_equals: expected false but got true
FAIL #4 Test that WebAssembly compilation succeeds (memory_init64.wast:94) assert_equals: expected false but got true
FAIL #5 Test that WebAssembly compilation succeeds (memory_init64.wast:138) assert_equals: expected false but got true
PASS #6 Test that WebAssembly compilation fails (memory_init64.wast:189)
PASS #7 Test that WebAssembly compilation fails (memory_init64.wast:195)
PASS #8 Test that WebAssembly compilation succeeds (memory_init64.wast:203)
PASS #9 Test that WebAssembly compilation succeeds (memory_init64.wast:211)
FAIL #10 Test that WebAssembly compilation succeeds (memory_init64.wast:219) assert_equals: expected false but got true
PASS #11 Test that WebAssembly compilation fails (memory_init64.wast:226)
PASS #12 Test that WebAssembly compilation fails (memory_init64.wast:232)
PASS #13 Test that WebAssembly compilation succeeds (memory_init64.wast:240)
PASS #14 Test that WebAssembly compilation succeeds (memory_init64.wast:248)
PASS #15 Test that WebAssembly compilation succeeds (memory_init64.wast:255)
PASS #16 Test that WebAssembly compilation succeeds (memory_init64.wast:262)
PASS #17 Test that WebAssembly compilation succeeds (memory_init64.wast:269)
PASS #18 Test that WebAssembly compilation succeeds (memory_init64.wast:276)
PASS #19 Test that WebAssembly compilation succeeds (memory_init64.wast:283)
PASS #20 Test that WebAssembly compilation succeeds (memory_init64.wast:290)
PASS #21 Test that WebAssembly compilation succeeds (memory_init64.wast:297)
PASS #22 Test that WebAssembly compilation succeeds (memory_init64.wast:304)
PASS #23 Test that WebAssembly compilation fails (memory_init64.wast:311)
PASS #24 Test that WebAssembly compilation fails (memory_init64.wast:319)
PASS #25 Test that WebAssembly compilation fails (memory_init64.wast:327)
PASS #26 Test that WebAssembly compilation fails (memory_init64.wast:335)
PASS #27 Test that WebAssembly compilation fails (memory_init64.wast:343)
PASS #28 Test that WebAssembly compilation fails (memory_init64.wast:351)
PASS #29 Test that WebAssembly compilation fails (memory_init64.wast:359)
PASS #30 Test that WebAssembly compilation fails (memory_init64.wast:367)
PASS #31 Test that WebAssembly compilation fails (memory_init64.wast:375)
PASS #32 Test that WebAssembly compilation fails (memory_init64.wast:383)
PASS #33 Test that WebAssembly compilation fails (memory_init64.wast:391)
PASS #34 Test that WebAssembly compilation fails (memory_init64.wast:399)
PASS #35 Test that WebAssembly compilation fails (memory_init64.wast:407)
PASS #36 Test that WebAssembly compilation fails (memory_init64.wast:415)
PASS #37 Test that WebAssembly compilation fails (memory_init64.wast:423)
PASS #38 Test that WebAssembly compilation fails (memory_init64.wast:431)
PASS #39 Test that WebAssembly compilation fails (memory_init64.wast:439)
PASS #40 Test that WebAssembly compilation fails (memory_init64.wast:447)
PASS #41 Test that WebAssembly compilation fails (memory_init64.wast:455)
PASS #42 Test that WebAssembly compilation fails (memory_init64.wast:463)
PASS #43 Test that WebAssembly compilation fails (memory_init64.wast:471)
PASS #44 Test that WebAssembly compilation fails (memory_init64.wast:479)
PASS #45 Test that WebAssembly compilation fails (memory_init64.wast:487)
PASS #46 Test that WebAssembly compilation fails (memory_init64.wast:495)
PASS #47 Test that WebAssembly compilation fails (memory_init64.wast:503)
PASS #48 Test that WebAssembly compilation fails (memory_init64.wast:511)
PASS #49 Test that WebAssembly compilation fails (memory_init64.wast:519)
PASS #50 Test that WebAssembly compilation fails (memory_init64.wast:527)
PASS #51 Test that WebAssembly compilation fails (memory_init64.wast:535)
PASS #52 Test that WebAssembly compilation fails (memory_init64.wast:543)
PASS #53 Test that WebAssembly compilation fails (memory_init64.wast:551)
PASS #54 Test that WebAssembly compilation fails (memory_init64.wast:559)
PASS #55 Test that WebAssembly compilation fails (memory_init64.wast:567)
PASS #56 Test that WebAssembly compilation fails (memory_init64.wast:575)
PASS #57 Test that WebAssembly compilation fails (memory_init64.wast:583)
PASS #58 Test that WebAssembly compilation fails (memory_init64.wast:591)
PASS #59 Test that WebAssembly compilation fails (memory_init64.wast:599)
PASS #60 Test that WebAssembly compilation fails (memory_init64.wast:607)
PASS #61 Test that WebAssembly compilation fails (memory_init64.wast:615)
PASS #62 Test that WebAssembly compilation fails (memory_init64.wast:623)
PASS #63 Test that WebAssembly compilation fails (memory_init64.wast:631)
PASS #64 Test that WebAssembly compilation fails (memory_init64.wast:639)
PASS #65 Test that WebAssembly compilation fails (memory_init64.wast:647)
PASS #66 Test that WebAssembly compilation fails (memory_init64.wast:655)
PASS #67 Test that WebAssembly compilation fails (memory_init64.wast:663)
PASS #68 Test that WebAssembly compilation fails (memory_init64.wast:671)
PASS #69 Test that WebAssembly compilation fails (memory_init64.wast:679)
PASS #70 Test that WebAssembly compilation fails (memory_init64.wast:687)
PASS #71 Test that WebAssembly compilation fails (memory_init64.wast:695)
PASS #72 Test that WebAssembly compilation fails (memory_init64.wast:703)
PASS #73 Test that WebAssembly compilation fails (memory_init64.wast:711)
PASS #74 Test that WebAssembly compilation fails (memory_init64.wast:719)
PASS #75 Test that WebAssembly compilation fails (memory_init64.wast:727)
PASS #76 Test that WebAssembly compilation fails (memory_init64.wast:735)
PASS #77 Test that WebAssembly compilation fails (memory_init64.wast:743)
PASS #78 Test that WebAssembly compilation fails (memory_init64.wast:751)
PASS #79 Test that WebAssembly compilation fails (memory_init64.wast:759)
PASS #80 Test that WebAssembly compilation fails (memory_init64.wast:767)
PASS #81 Test that WebAssembly compilation fails (memory_init64.wast:775)
PASS #82 Test that WebAssembly compilation fails (memory_init64.wast:783)
PASS #83 Test that WebAssembly compilation fails (memory_init64.wast:791)
PASS #84 Test that WebAssembly compilation fails (memory_init64.wast:799)
PASS #85 Test that WebAssembly compilation fails (memory_init64.wast:807)
PASS #86 Test that WebAssembly compilation succeeds (memory_init64.wast:815)
PASS #87 Test that WebAssembly compilation succeeds (memory_init64.wast:838)
PASS #88 Test that WebAssembly compilation succeeds (memory_init64.wast:861)
PASS #89 Test that WebAssembly compilation succeeds (memory_init64.wast:884)
PASS #90 Test that WebAssembly compilation succeeds (memory_init64.wast:907)
PASS #91 Test that WebAssembly compilation succeeds (memory_init64.wast:930)
PASS #92 Test that WebAssembly compilation succeeds (memory_init64.wast:954)
PASS #93 Reinitialize the default imports
FAIL #94 Test that WebAssembly compilation succeeds (memory_init64.wast:6) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 114: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #95 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #96 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:10:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #97 Test that a WebAssembly code returns a specific result (memory_init64.wast:19) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:13:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #98 Test that a WebAssembly code returns a specific result (memory_init64.wast:20) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #99 Test that a WebAssembly code returns a specific result (memory_init64.wast:21) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #100 Test that a WebAssembly code returns a specific result (memory_init64.wast:22) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #101 Test that a WebAssembly code returns a specific result (memory_init64.wast:23) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #102 Test that a WebAssembly code returns a specific result (memory_init64.wast:24) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #103 Test that a WebAssembly code returns a specific result (memory_init64.wast:25) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #104 Test that a WebAssembly code returns a specific result (memory_init64.wast:26) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #105 Test that a WebAssembly code returns a specific result (memory_init64.wast:27) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #106 Test that a WebAssembly code returns a specific result (memory_init64.wast:28) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #107 Test that a WebAssembly code returns a specific result (memory_init64.wast:29) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #108 Test that a WebAssembly code returns a specific result (memory_init64.wast:30) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #109 Test that a WebAssembly code returns a specific result (memory_init64.wast:31) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #110 Test that a WebAssembly code returns a specific result (memory_init64.wast:32) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #111 Test that a WebAssembly code returns a specific result (memory_init64.wast:33) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #112 Test that a WebAssembly code returns a specific result (memory_init64.wast:34) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:58:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #113 Test that a WebAssembly code returns a specific result (memory_init64.wast:35) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #114 Test that a WebAssembly code returns a specific result (memory_init64.wast:36) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #115 Test that a WebAssembly code returns a specific result (memory_init64.wast:37) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #116 Test that a WebAssembly code returns a specific result (memory_init64.wast:38) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #117 Test that a WebAssembly code returns a specific result (memory_init64.wast:39) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #118 Test that a WebAssembly code returns a specific result (memory_init64.wast:40) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #119 Test that a WebAssembly code returns a specific result (memory_init64.wast:41) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:79:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #120 Test that a WebAssembly code returns a specific result (memory_init64.wast:42) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #121 Test that a WebAssembly code returns a specific result (memory_init64.wast:43) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #122 Test that a WebAssembly code returns a specific result (memory_init64.wast:44) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #123 Test that a WebAssembly code returns a specific result (memory_init64.wast:45) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #124 Test that a WebAssembly code returns a specific result (memory_init64.wast:46) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #125 Test that a WebAssembly code returns a specific result (memory_init64.wast:47) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:97:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #126 Test that a WebAssembly code returns a specific result (memory_init64.wast:48) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:100:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #127 Test that WebAssembly compilation succeeds (memory_init64.wast:50) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 130: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #128 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:106:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #129 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:109:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #130 Test that a WebAssembly code returns a specific result (memory_init64.wast:63) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #131 Test that a WebAssembly code returns a specific result (memory_init64.wast:64) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #132 Test that a WebAssembly code returns a specific result (memory_init64.wast:65) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #133 Test that a WebAssembly code returns a specific result (memory_init64.wast:66) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:121:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #134 Test that a WebAssembly code returns a specific result (memory_init64.wast:67) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:124:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #135 Test that a WebAssembly code returns a specific result (memory_init64.wast:68) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:127:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #136 Test that a WebAssembly code returns a specific result (memory_init64.wast:69) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:130:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #137 Test that a WebAssembly code returns a specific result (memory_init64.wast:70) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:133:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #138 Test that a WebAssembly code returns a specific result (memory_init64.wast:71) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:136:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #139 Test that a WebAssembly code returns a specific result (memory_init64.wast:72) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:139:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #140 Test that a WebAssembly code returns a specific result (memory_init64.wast:73) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:142:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #141 Test that a WebAssembly code returns a specific result (memory_init64.wast:74) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:145:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #142 Test that a WebAssembly code returns a specific result (memory_init64.wast:75) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:148:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #143 Test that a WebAssembly code returns a specific result (memory_init64.wast:76) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:151:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #144 Test that a WebAssembly code returns a specific result (memory_init64.wast:77) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:154:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #145 Test that a WebAssembly code returns a specific result (memory_init64.wast:78) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:157:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #146 Test that a WebAssembly code returns a specific result (memory_init64.wast:79) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:160:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #147 Test that a WebAssembly code returns a specific result (memory_init64.wast:80) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:163:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #148 Test that a WebAssembly code returns a specific result (memory_init64.wast:81) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:166:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #149 Test that a WebAssembly code returns a specific result (memory_init64.wast:82) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:169:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #150 Test that a WebAssembly code returns a specific result (memory_init64.wast:83) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:172:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #151 Test that a WebAssembly code returns a specific result (memory_init64.wast:84) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:175:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #152 Test that a WebAssembly code returns a specific result (memory_init64.wast:85) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:178:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #153 Test that a WebAssembly code returns a specific result (memory_init64.wast:86) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:181:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #154 Test that a WebAssembly code returns a specific result (memory_init64.wast:87) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:184:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #155 Test that a WebAssembly code returns a specific result (memory_init64.wast:88) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:187:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #156 Test that a WebAssembly code returns a specific result (memory_init64.wast:89) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:190:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #157 Test that a WebAssembly code returns a specific result (memory_init64.wast:90) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:193:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #158 Test that a WebAssembly code returns a specific result (memory_init64.wast:91) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:196:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #159 Test that a WebAssembly code returns a specific result (memory_init64.wast:92) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:199:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #160 Test that WebAssembly compilation succeeds (memory_init64.wast:94) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 130: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #161 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:205:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #162 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:208:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #163 Test that a WebAssembly code returns a specific result (memory_init64.wast:107) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:211:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #164 Test that a WebAssembly code returns a specific result (memory_init64.wast:108) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:214:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #165 Test that a WebAssembly code returns a specific result (memory_init64.wast:109) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:217:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #166 Test that a WebAssembly code returns a specific result (memory_init64.wast:110) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:220:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #167 Test that a WebAssembly code returns a specific result (memory_init64.wast:111) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:223:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #168 Test that a WebAssembly code returns a specific result (memory_init64.wast:112) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:226:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #169 Test that a WebAssembly code returns a specific result (memory_init64.wast:113) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:229:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #170 Test that a WebAssembly code returns a specific result (memory_init64.wast:114) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:232:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #171 Test that a WebAssembly code returns a specific result (memory_init64.wast:115) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:235:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #172 Test that a WebAssembly code returns a specific result (memory_init64.wast:116) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:238:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #173 Test that a WebAssembly code returns a specific result (memory_init64.wast:117) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:241:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #174 Test that a WebAssembly code returns a specific result (memory_init64.wast:118) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:244:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #175 Test that a WebAssembly code returns a specific result (memory_init64.wast:119) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:247:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #176 Test that a WebAssembly code returns a specific result (memory_init64.wast:120) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:250:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #177 Test that a WebAssembly code returns a specific result (memory_init64.wast:121) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:253:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #178 Test that a WebAssembly code returns a specific result (memory_init64.wast:122) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:256:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #179 Test that a WebAssembly code returns a specific result (memory_init64.wast:123) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:259:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #180 Test that a WebAssembly code returns a specific result (memory_init64.wast:124) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:262:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #181 Test that a WebAssembly code returns a specific result (memory_init64.wast:125) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:265:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #182 Test that a WebAssembly code returns a specific result (memory_init64.wast:126) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:268:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #183 Test that a WebAssembly code returns a specific result (memory_init64.wast:127) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:271:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #184 Test that a WebAssembly code returns a specific result (memory_init64.wast:128) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:274:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #185 Test that a WebAssembly code returns a specific result (memory_init64.wast:129) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:277:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #186 Test that a WebAssembly code returns a specific result (memory_init64.wast:130) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:280:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #187 Test that a WebAssembly code returns a specific result (memory_init64.wast:131) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:283:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #188 Test that a WebAssembly code returns a specific result (memory_init64.wast:132) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:286:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #189 Test that a WebAssembly code returns a specific result (memory_init64.wast:133) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:289:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #190 Test that a WebAssembly code returns a specific result (memory_init64.wast:134) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:292:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #191 Test that a WebAssembly code returns a specific result (memory_init64.wast:135) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:295:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #192 Test that a WebAssembly code returns a specific result (memory_init64.wast:136) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:298:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #193 Test that WebAssembly compilation succeeds (memory_init64.wast:138) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 196: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #194 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:304:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #195 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:307:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #196 Test that a WebAssembly code returns a specific result (memory_init64.wast:159) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:310:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #197 Test that a WebAssembly code returns a specific result (memory_init64.wast:160) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:313:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #198 Test that a WebAssembly code returns a specific result (memory_init64.wast:161) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:316:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #199 Test that a WebAssembly code returns a specific result (memory_init64.wast:162) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:319:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #200 Test that a WebAssembly code returns a specific result (memory_init64.wast:163) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:322:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #201 Test that a WebAssembly code returns a specific result (memory_init64.wast:164) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:325:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #202 Test that a WebAssembly code returns a specific result (memory_init64.wast:165) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:328:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #203 Test that a WebAssembly code returns a specific result (memory_init64.wast:166) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:331:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #204 Test that a WebAssembly code returns a specific result (memory_init64.wast:167) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:334:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #205 Test that a WebAssembly code returns a specific result (memory_init64.wast:168) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:337:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #206 Test that a WebAssembly code returns a specific result (memory_init64.wast:169) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:340:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #207 Test that a WebAssembly code returns a specific result (memory_init64.wast:170) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:343:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #208 Test that a WebAssembly code returns a specific result (memory_init64.wast:171) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:346:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #209 Test that a WebAssembly code returns a specific result (memory_init64.wast:172) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:349:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #210 Test that a WebAssembly code returns a specific result (memory_init64.wast:173) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:352:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #211 Test that a WebAssembly code returns a specific result (memory_init64.wast:174) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:355:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #212 Test that a WebAssembly code returns a specific result (memory_init64.wast:175) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:358:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #213 Test that a WebAssembly code returns a specific result (memory_init64.wast:176) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:361:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #214 Test that a WebAssembly code returns a specific result (memory_init64.wast:177) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:364:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #215 Test that a WebAssembly code returns a specific result (memory_init64.wast:178) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:367:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #216 Test that a WebAssembly code returns a specific result (memory_init64.wast:179) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:370:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #217 Test that a WebAssembly code returns a specific result (memory_init64.wast:180) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:373:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #218 Test that a WebAssembly code returns a specific result (memory_init64.wast:181) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:376:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #219 Test that a WebAssembly code returns a specific result (memory_init64.wast:182) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:379:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #220 Test that a WebAssembly code returns a specific result (memory_init64.wast:183) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:382:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #221 Test that a WebAssembly code returns a specific result (memory_init64.wast:184) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:385:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #222 Test that a WebAssembly code returns a specific result (memory_init64.wast:185) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:388:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #223 Test that a WebAssembly code returns a specific result (memory_init64.wast:186) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:391:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #224 Test that a WebAssembly code returns a specific result (memory_init64.wast:187) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:394:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #225 Test that a WebAssembly code returns a specific result (memory_init64.wast:188) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:397:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
PASS #226 Test that WebAssembly compilation fails (memory_init64.wast:189)
PASS #227 Test that WebAssembly compilation fails (memory_init64.wast:195)
PASS #228 Test that WebAssembly compilation succeeds (memory_init64.wast:203)
PASS #229 Test that WebAssembly instantiation succeeds (memory_init64.wast:203)
PASS #230 Run a WebAssembly test without special assertions (memory_init64.wast:209)
PASS #231 Test that WebAssembly compilation succeeds (memory_init64.wast:211)
PASS #232 Test that WebAssembly instantiation succeeds (memory_init64.wast:211)
PASS #233 Test that a WebAssembly code traps (memory_init64.wast:217)
FAIL #234 Test that WebAssembly compilation succeeds (memory_init64.wast:219) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 92: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #235 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:427:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
FAIL #236 Test that a WebAssembly code traps (memory_init64.wast:224) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
memory_init64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:430:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_init64.wast.js:795:3 expected true got false
PASS #237 Test that WebAssembly compilation fails (memory_init64.wast:226)
PASS #238 Test that WebAssembly compilation fails (memory_init64.wast:232)
PASS #239 Test that WebAssembly compilation succeeds (memory_init64.wast:240)
PASS #240 Test that WebAssembly instantiation succeeds (memory_init64.wast:240)
PASS #241 Run a WebAssembly test without special assertions (memory_init64.wast:246)
PASS #242 Test that WebAssembly compilation succeeds (memory_init64.wast:248)
PASS #243 Test that WebAssembly instantiation succeeds (memory_init64.wast:248)
PASS #244 Test that a WebAssembly code traps (memory_init64.wast:253)
PASS #245 Test that WebAssembly compilation succeeds (memory_init64.wast:255)
PASS #246 Test that WebAssembly instantiation succeeds (memory_init64.wast:255)
PASS #247 Test that a WebAssembly code traps (memory_init64.wast:260)
PASS #248 Test that WebAssembly compilation succeeds (memory_init64.wast:262)
PASS #249 Test that WebAssembly instantiation succeeds (memory_init64.wast:262)
PASS #250 Test that a WebAssembly code traps (memory_init64.wast:267)
PASS #251 Test that WebAssembly compilation succeeds (memory_init64.wast:269)
PASS #252 Test that WebAssembly instantiation succeeds (memory_init64.wast:269)
PASS #253 Test that a WebAssembly code traps (memory_init64.wast:274)
PASS #254 Test that WebAssembly compilation succeeds (memory_init64.wast:276)
PASS #255 Test that WebAssembly instantiation succeeds (memory_init64.wast:276)
PASS #256 Run a WebAssembly test without special assertions (memory_init64.wast:281)
PASS #257 Test that WebAssembly compilation succeeds (memory_init64.wast:283)
PASS #258 Test that WebAssembly instantiation succeeds (memory_init64.wast:283)
PASS #259 Test that a WebAssembly code traps (memory_init64.wast:288)
PASS #260 Test that WebAssembly compilation succeeds (memory_init64.wast:290)
PASS #261 Test that WebAssembly instantiation succeeds (memory_init64.wast:290)
PASS #262 Run a WebAssembly test without special assertions (memory_init64.wast:295)
PASS #263 Test that WebAssembly compilation succeeds (memory_init64.wast:297)
PASS #264 Test that WebAssembly instantiation succeeds (memory_init64.wast:297)
PASS #265 Run a WebAssembly test without special assertions (memory_init64.wast:302)
PASS #266 Test that WebAssembly compilation succeeds (memory_init64.wast:304)
PASS #267 Test that WebAssembly instantiation succeeds (memory_init64.wast:304)
PASS #268 Test that a WebAssembly code traps (memory_init64.wast:309)
PASS #269 Test that WebAssembly compilation fails (memory_init64.wast:311)
PASS #270 Test that WebAssembly compilation fails (memory_init64.wast:319)
PASS #271 Test that WebAssembly compilation fails (memory_init64.wast:327)
PASS #272 Test that WebAssembly compilation fails (memory_init64.wast:335)
PASS #273 Test that WebAssembly compilation fails (memory_init64.wast:343)
PASS #274 Test that WebAssembly compilation fails (memory_init64.wast:351)
PASS #275 Test that WebAssembly compilation fails (memory_init64.wast:359)
PASS #276 Test that WebAssembly compilation fails (memory_init64.wast:367)
PASS #277 Test that WebAssembly compilation fails (memory_init64.wast:375)
PASS #278 Test that WebAssembly compilation fails (memory_init64.wast:383)
PASS #279 Test that WebAssembly compilation fails (memory_init64.wast:391)
PASS #280 Test that WebAssembly compilation fails (memory_init64.wast:399)
PASS #281 Test that WebAssembly compilation fails (memory_init64.wast:407)
PASS #282 Test that WebAssembly compilation fails (memory_init64.wast:415)
PASS #283 Test that WebAssembly compilation fails (memory_init64.wast:423)
PASS #284 Test that WebAssembly compilation fails (memory_init64.wast:431)
PASS #285 Test that WebAssembly compilation fails (memory_init64.wast:439)
PASS #286 Test that WebAssembly compilation fails (memory_init64.wast:447)
PASS #287 Test that WebAssembly compilation fails (memory_init64.wast:455)
PASS #288 Test that WebAssembly compilation fails (memory_init64.wast:463)
PASS #289 Test that WebAssembly compilation fails (memory_init64.wast:471)
PASS #290 Test that WebAssembly compilation fails (memory_init64.wast:479)
PASS #291 Test that WebAssembly compilation fails (memory_init64.wast:487)
PASS #292 Test that WebAssembly compilation fails (memory_init64.wast:495)
PASS #293 Test that WebAssembly compilation fails (memory_init64.wast:503)
PASS #294 Test that WebAssembly compilation fails (memory_init64.wast:511)
PASS #295 Test that WebAssembly compilation fails (memory_init64.wast:519)
PASS #296 Test that WebAssembly compilation fails (memory_init64.wast:527)
PASS #297 Test that WebAssembly compilation fails (memory_init64.wast:535)
PASS #298 Test that WebAssembly compilation fails (memory_init64.wast:543)
PASS #299 Test that WebAssembly compilation fails (memory_init64.wast:551)
PASS #300 Test that WebAssembly compilation fails (memory_init64.wast:559)
PASS #301 Test that WebAssembly compilation fails (memory_init64.wast:567)
PASS #302 Test that WebAssembly compilation fails (memory_init64.wast:575)
PASS #303 Test that WebAssembly compilation fails (memory_init64.wast:583)
PASS #304 Test that WebAssembly compilation fails (memory_init64.wast:591)
PASS #305 Test that WebAssembly compilation fails (memory_init64.wast:599)
PASS #306 Test that WebAssembly compilation fails (memory_init64.wast:607)
PASS #307 Test that WebAssembly compilation fails (memory_init64.wast:615)
PASS #308 Test that WebAssembly compilation fails (memory_init64.wast:623)
PASS #309 Test that WebAssembly compilation fails (memory_init64.wast:631)
PASS #310 Test that WebAssembly compilation fails (memory_init64.wast:639)
PASS #311 Test that WebAssembly compilation fails (memory_init64.wast:647)
PASS #312 Test that WebAssembly compilation fails (memory_init64.wast:655)
PASS #313 Test that WebAssembly compilation fails (memory_init64.wast:663)
PASS #314 Test that WebAssembly compilation fails (memory_init64.wast:671)
PASS #315 Test that WebAssembly compilation fails (memory_init64.wast:679)
PASS #316 Test that WebAssembly compilation fails (memory_init64.wast:687)
PASS #317 Test that WebAssembly compilation fails (memory_init64.wast:695)
PASS #318 Test that WebAssembly compilation fails (memory_init64.wast:703)
PASS #319 Test that WebAssembly compilation fails (memory_init64.wast:711)
PASS #320 Test that WebAssembly compilation fails (memory_init64.wast:719)
PASS #321 Test that WebAssembly compilation fails (memory_init64.wast:727)
PASS #322 Test that WebAssembly compilation fails (memory_init64.wast:735)
PASS #323 Test that WebAssembly compilation fails (memory_init64.wast:743)
PASS #324 Test that WebAssembly compilation fails (memory_init64.wast:751)
PASS #325 Test that WebAssembly compilation fails (memory_init64.wast:759)
PASS #326 Test that WebAssembly compilation fails (memory_init64.wast:767)
PASS #327 Test that WebAssembly compilation fails (memory_init64.wast:775)
PASS #328 Test that WebAssembly compilation fails (memory_init64.wast:783)
PASS #329 Test that WebAssembly compilation fails (memory_init64.wast:791)
PASS #330 Test that WebAssembly compilation fails (memory_init64.wast:799)
PASS #331 Test that WebAssembly compilation fails (memory_init64.wast:807)
PASS #332 Test that WebAssembly compilation succeeds (memory_init64.wast:815)
PASS #333 Test that WebAssembly instantiation succeeds (memory_init64.wast:815)
PASS #334 Test that a WebAssembly code traps (memory_init64.wast:833)
PASS #335 Test that a WebAssembly code returns a specific result (memory_init64.wast:836)
PASS #336 Test that WebAssembly compilation succeeds (memory_init64.wast:838)
PASS #337 Test that WebAssembly instantiation succeeds (memory_init64.wast:838)
PASS #338 Test that a WebAssembly code traps (memory_init64.wast:856)
PASS #339 Test that a WebAssembly code returns a specific result (memory_init64.wast:859)
PASS #340 Test that WebAssembly compilation succeeds (memory_init64.wast:861)
PASS #341 Test that WebAssembly instantiation succeeds (memory_init64.wast:861)
PASS #342 Test that a WebAssembly code traps (memory_init64.wast:879)
PASS #343 Test that a WebAssembly code returns a specific result (memory_init64.wast:882)
PASS #344 Test that WebAssembly compilation succeeds (memory_init64.wast:884)
PASS #345 Test that WebAssembly instantiation succeeds (memory_init64.wast:884)
PASS #346 Test that a WebAssembly code traps (memory_init64.wast:902)
PASS #347 Test that a WebAssembly code returns a specific result (memory_init64.wast:905)
PASS #348 Test that WebAssembly compilation succeeds (memory_init64.wast:907)
PASS #349 Test that WebAssembly instantiation succeeds (memory_init64.wast:907)
PASS #350 Test that a WebAssembly code traps (memory_init64.wast:925)
PASS #351 Test that a WebAssembly code returns a specific result (memory_init64.wast:928)
PASS #352 Test that WebAssembly compilation succeeds (memory_init64.wast:930)
PASS #353 Test that WebAssembly instantiation succeeds (memory_init64.wast:930)
PASS #354 Test that a WebAssembly code traps (memory_init64.wast:948)
PASS #355 Test that a WebAssembly code returns a specific result (memory_init64.wast:951)
PASS #356 Test that WebAssembly compilation succeeds (memory_init64.wast:954)
PASS #357 Test that WebAssembly instantiation succeeds (memory_init64.wast:954)
